Waiting for Tests to End

+++ PerfTester-Xamarin/PerfTester-Xamarin/Models/Helpers/Dog.cs
@@ -6,7 +6,7 @@ namespace PerfTesterXamarin.Models.Helpers
 {
     public class Dog : RealmObject
     {
-        public Int16 Id { get; set; }
+        public Int32 Id { get; set; }
         public string Name { get; set; }
         public int Age { get; set; }
         public Person Owner { get; set; }
