+++ PerfTester-Xamarin/PerfTester-Xamarin/Models/Helpers/Person.cs
@@ -6,7 +6,7 @@ namespace PerfTesterXamarin.Models.Helpers
 {
     public class Person : RealmObject
     {
-        public Int16 Id { get; set; }
+        public Int32 Id { get; set; }
         public string Name { get; set; }
         public IList<Dog> Dogs { get; }
     }
